requirements.txt
scipy==1.3.1
opencv_python==4.2.0.34
Shapely==1.7.0
matplotlib==3.1.1
numpy==1.17.2
mmcv==1.2.1
easydict==1.9
numba==0.51.0
Pillow==8.1.0
PyYAML==5.3.1
tensorboardX==2.1
